+++ app/models/forms/bulk_upload_form/checking_file.rb
@@ -1,22 +1,23 @@
 module Forms
-  module BulkUploadLettings
+  module BulkUploadForm
     class CheckingFile
       include ActiveModel::Model
       include ActiveModel::Attributes
       include Rails.application.routes.url_helpers
 
+      attribute :log_type
       attribute :year, :integer
       attribute :organisation_id, :integer
 
       def view_path
-        "bulk_upload_lettings_logs/forms/checking_file"
+        "bulk_upload_#{log_type}_logs/forms/checking_file"
       end
 
       def back_path
         if organisation_id.present?
-          lettings_logs_organisation_path(organisation_id)
+          send("#{log_type}_logs_organisation_path", organisation_id)
         else
-          bulk_upload_lettings_log_path(id: "start")
+          send("bulk_upload_#{log_type}_log_path", id: "start")
         end
       end
